requirements.txt
anaconda==0.0.1.1
appdirs==1.4.3
asgiref==3.2.3
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cssselect==1.1.0
cycler==0.10.0
django==3.0.1
django-phonenumber-field==5.1.0
django-suit==0.3a1
fake-useragent==0.1.11
Faker==4.1.1
Flask==1.1.2
Flask-SQLAlchemy==2.4.3
gunicorn==20.1.0
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
lxml==4.5.0
MarkupSafe==1.1.1
matplotlib==3.2.0
mccabe==0.6.1
more-itertools==8.1.0
mysqlclient==1.4.6
numpy==1.18.1
packaging==20.0
pandas==1.2.2
parse==1.14.0
phonenumbers==8.12.22
php==1.2.1
Pillow==8.2.0
pluggy==0.13.1
py==1.8.1
pyee==7.0.1
pylint==2.4.4
pyparsing==2.4.6
pyppeteer==0.0.25
PyQt5==5.14.1
PyQt5-sip==12.7.0
pyquery==1.4.1
pytest==5.3.3
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
requests-html==0.10.0
six==1.13.0
soupsieve==1.9.5
SQLAlchemy==1.3.18
sqlparse==0.3.0
text-unidecode==1.3
tqdm==4.42.1
urllib3==1.25.8
virtualenv==16.7.9
virtualenvwrapper-win==1.2.5
w3lib==1.21.0
wcwidth==0.1.8
websockets==8.1
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.11.2
